CoinModel and script added

+++ CarEnergyDrifting3D-URP/Assets/MainScripts/LevelController.cs
@@ -4,12 +4,12 @@ using UnityEngine;
 
 public class LevelController : MonoBehaviour
 {
-    public static float currentLevel, levelSave;
+    public static int currentLevel, levelSave;
 
     void Start()
     {
      //   currentLevel = PlayerPrefs.GetFloat("level_main", 0);
-        Debug.Log("CurrentLevel: " + PlayerPrefs.GetFloat("level_main"));
+        Debug.Log("CurrentLevel: " + PlayerPrefs.GetInt("level_main"));
     }
     
     void Update()
